Move to long int so we can store hashed values there when needed.

+++ src/list.h
@@ -4,7 +4,7 @@
 struct node {
   bstring name;
   bstring content;
-  int level;
+  long level;
   struct node *next;
 };
